Move source files into a src directory
